requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
boto3==1.34.0
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
whitenoise==6.6.0
gunicorn==21.2.0 